+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to the "vscode-kusto-workbench" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [4.7.4] - 2026.06.28
+## [4.7.4] - 2026.06.30
 
 * Improved 'file open' performance.
 * Bugs
@@ -11,6 +11,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     * Connection Manager would in certain cases require you to scroll before it would display the databases of the selected cluster (scroll view bug)
     * Adding / removing from favorites while files were open meant the files had to be re-opened before they saw the changes.
     * Renaming files but keeping most of the filename the same would cause weird situations where both files would appear side-by-side.
+    * Connection Manager: the text across a single row was not vertically aligned. 
+    * Connection Manager: fixed some scrollbar bugs. 
 
 ## [4.7.3] - 2026.06.08